refactor(query): Split planner/index_selection.rs into submodules
Split 1,742-line index_selection.rs into 6 focused files:
- types.rs: IndexScanParams, ScanValueFilter, IndexScanType, FieldCondition, ConditionValue
- values.rs: Value conversion and normalization helpers
- conditions.rs: Condition analysis, scoring, and index compatibility checks
- filter_to_scan.rs: Core filter-to-index-scan conversion and OR handling
- tests.rs: All 31 tests
- mod.rs: Module declarations and re-exports
